pyproject.toml
[project]
name = "vision-ml-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dagshub>=0.3.0",
    "matplotlib>=3.8.0",
    "mlflow>=2.10.0",
    "numpy>=1.24.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.17.0",
    "opencv-python>=4.9.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "plotly>=5.18.0",
    "psutil>=5.9.0",
    "pytest>=8.0.0",
    "pyyaml>=6.0",
    "rfdetr>=1.4.3",
    "streamlit>=1.28.0",
    "supervision>=0.24.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "tqdm>=4.66.0",
    "trackers>=2.2.0",
    "ultralytics>=8.3.0",
]
